feat(#119): Realiza testes com home, navbar, onboarding, profile e welcome
