--- lib/ui/edit_password/di/di.dart
@@ -1,9 +1,0 @@
-import 'package:aranduapp/ui/edit_password/service/edit_password_service.dart';
-import 'package:aranduapp/ui/edit_password/viewmodel/edit_password_viewmodel.dart';
-import 'package:get_it/get_it.dart';
-
-void setupPasswordDI(){
-
-  GetIt.instance.registerFactory(() => EditPasswordViewModel());
-  GetIt.instance.registerSingleton(() => EditPasswordService());
-}
