+++ test/ui/navbar/view/navbar_view_test.mocks.dart
@@ -1,13 +1,12 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in aranduapp/test/ui/onboarding/view/onboarding_test.dart.
+// in aranduapp/test/ui/navbar/view/navbar_view_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:ui' as _i4;
 
-import 'package:aranduapp/ui/onboarding/viewmodel/onboarding_viewmodel.dart'
-    as _i3;
-import 'package:flutter/material.dart' as _i2;
+import 'package:aranduapp/core/state/command.dart' as _i2;
+import 'package:aranduapp/ui/navbar/viewmodel/navbar_viewmodel.dart' as _i3;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -23,9 +22,8 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakePageController_0 extends _i1.SmartFake
-    implements _i2.PageController {
-  _FakePageController_0(
+class _FakeCommand0_0<T> extends _i1.SmartFake implements _i2.Command0<T> {
+  _FakeCommand0_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -34,39 +32,38 @@ class _FakePageController_0 extends _i1.SmartFake
         );
 }
 
-/// A class which mocks [OnboardingViewModel].
+/// A class which mocks [NavbarViewModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockOnboardingViewModel extends _i1.Mock
-    implements _i3.OnboardingViewModel {
+class MockNavbarViewModel extends _i1.Mock implements _i3.NavbarViewModel {
   @override
-  _i2.PageController get pageController => (super.noSuchMethod(
-        Invocation.getter(#pageController),
-        returnValue: _FakePageController_0(
+  _i2.Command0<void> get navbarCommand => (super.noSuchMethod(
+        Invocation.getter(#navbarCommand),
+        returnValue: _FakeCommand0_0<void>(
           this,
-          Invocation.getter(#pageController),
+          Invocation.getter(#navbarCommand),
         ),
-        returnValueForMissingStub: _FakePageController_0(
+        returnValueForMissingStub: _FakeCommand0_0<void>(
           this,
-          Invocation.getter(#pageController),
+          Invocation.getter(#navbarCommand),
         ),
-      ) as _i2.PageController);
+      ) as _i2.Command0<void>);
 
   @override
-  int get currentPage => (super.noSuchMethod(
-        Invocation.getter(#currentPage),
-        returnValue: 0,
-        returnValueForMissingStub: 0,
-      ) as int);
-
-  @override
-  set currentPage(int? _currentPage) => super.noSuchMethod(
+  set navbarCommand(_i2.Command0<void>? _navbarCommand) => super.noSuchMethod(
         Invocation.setter(
-          #currentPage,
-          _currentPage,
+          #navbarCommand,
+          _navbarCommand,
         ),
         returnValueForMissingStub: null,
       );
+
+  @override
+  int get selectedIndex => (super.noSuchMethod(
+        Invocation.getter(#selectedIndex),
+        returnValue: 0,
+        returnValueForMissingStub: 0,
+      ) as int);
 
   @override
   bool get hasListeners => (super.noSuchMethod(
@@ -76,28 +73,10 @@ class MockOnboardingViewModel extends _i1.Mock
       ) as bool);
 
   @override
-  void goToNextPage() => super.noSuchMethod(
+  void changeTab(int? index) => super.noSuchMethod(
         Invocation.method(
-          #goToNextPage,
-          [],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void updateCurrentPage(int? page) => super.noSuchMethod(
-        Invocation.method(
-          #updateCurrentPage,
-          [page],
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  void navigateToNextPage(_i2.BuildContext? context) => super.noSuchMethod(
-        Invocation.method(
-          #navigateToNextPage,
-          [context],
+          #changeTab,
+          [index],
         ),
         returnValueForMissingStub: null,
       );
