+++ test/ui/onboarding/view/onboarding_view_test.mocks.dart
@@ -1,12 +1,13 @@
 // Mocks generated by Mockito 5.4.4 from annotations
-// in aranduapp/test/ui/navbar/view/navbar_test.dart.
+// in aranduapp/test/ui/onboarding/view/onboarding_view_test.dart.
 // Do not manually edit this file.
 
 // ignore_for_file: no_leading_underscores_for_library_prefixes
 import 'dart:ui' as _i4;
 
-import 'package:aranduapp/core/state/command.dart' as _i2;
-import 'package:aranduapp/ui/navbar/viewmodel/navbar_viewmodel.dart' as _i3;
+import 'package:aranduapp/ui/onboarding/viewmodel/onboarding_viewmodel.dart'
+    as _i3;
+import 'package:flutter/material.dart' as _i2;
 import 'package:mockito/mockito.dart' as _i1;
 
 // ignore_for_file: type=lint
@@ -22,8 +23,9 @@ import 'package:mockito/mockito.dart' as _i1;
 // ignore_for_file: camel_case_types
 // ignore_for_file: subtype_of_sealed_class
 
-class _FakeCommand0_0<T> extends _i1.SmartFake implements _i2.Command0<T> {
-  _FakeCommand0_0(
+class _FakePageController_0 extends _i1.SmartFake
+    implements _i2.PageController {
+  _FakePageController_0(
     Object parent,
     Invocation parentInvocation,
   ) : super(
@@ -32,38 +34,39 @@ class _FakeCommand0_0<T> extends _i1.SmartFake implements _i2.Command0<T> {
         );
 }
 
-/// A class which mocks [NavbarViewModel].
+/// A class which mocks [OnboardingViewModel].
 ///
 /// See the documentation for Mockito's code generation for more information.
-class MockNavbarViewModel extends _i1.Mock implements _i3.NavbarViewModel {
+class MockOnboardingViewModel extends _i1.Mock
+    implements _i3.OnboardingViewModel {
   @override
-  _i2.Command0<void> get navbarCommand => (super.noSuchMethod(
-        Invocation.getter(#navbarCommand),
-        returnValue: _FakeCommand0_0<void>(
+  _i2.PageController get pageController => (super.noSuchMethod(
+        Invocation.getter(#pageController),
+        returnValue: _FakePageController_0(
           this,
-          Invocation.getter(#navbarCommand),
+          Invocation.getter(#pageController),
         ),
-        returnValueForMissingStub: _FakeCommand0_0<void>(
+        returnValueForMissingStub: _FakePageController_0(
           this,
-          Invocation.getter(#navbarCommand),
+          Invocation.getter(#pageController),
         ),
-      ) as _i2.Command0<void>);
+      ) as _i2.PageController);
 
   @override
-  set navbarCommand(_i2.Command0<void>? _navbarCommand) => super.noSuchMethod(
-        Invocation.setter(
-          #navbarCommand,
-          _navbarCommand,
-        ),
-        returnValueForMissingStub: null,
-      );
-
-  @override
-  int get selectedIndex => (super.noSuchMethod(
-        Invocation.getter(#selectedIndex),
+  int get currentPage => (super.noSuchMethod(
+        Invocation.getter(#currentPage),
         returnValue: 0,
         returnValueForMissingStub: 0,
       ) as int);
+
+  @override
+  set currentPage(int? _currentPage) => super.noSuchMethod(
+        Invocation.setter(
+          #currentPage,
+          _currentPage,
+        ),
+        returnValueForMissingStub: null,
+      );
 
   @override
   bool get hasListeners => (super.noSuchMethod(
@@ -73,10 +76,28 @@ class MockNavbarViewModel extends _i1.Mock implements _i3.NavbarViewModel {
       ) as bool);
 
   @override
-  void changeTab(int? index) => super.noSuchMethod(
+  void goToNextPage() => super.noSuchMethod(
         Invocation.method(
-          #changeTab,
-          [index],
+          #goToNextPage,
+          [],
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void updateCurrentPage(int? page) => super.noSuchMethod(
+        Invocation.method(
+          #updateCurrentPage,
+          [page],
+        ),
+        returnValueForMissingStub: null,
+      );
+
+  @override
+  void navigateToNextPage(_i2.BuildContext? context) => super.noSuchMethod(
+        Invocation.method(
+          #navigateToNextPage,
+          [context],
         ),
         returnValueForMissingStub: null,
       );
