--- test/ui/onboarding/viewmodel/onboarding_viewmodel_test.dart
@@ -1,29 +1,0 @@
-//  testWidgets('navegation', (WidgetTester tester) async {
-//    await tester.pumpWidget(createScreen());
-
-//    expect(find.text('Bem-Vindo(a)!'), findsOneWidget);
-
-    // Clica no botão 'Avançar' para navegar para o próximo passo
-//    await tester.tap(find.text('Avançar'));
-//    await tester.pumpAndSettle(); // Aguarda a animação/página carregar
-
-    // Verifica se o segundo título está visível após clicar no botão
-//    expect(find.text('o que você encontra?'), findsOneWidget);
-
-    // Clica novamente no botão 'Avançar' para ir para o próximo passo
- //   await tester.tap(find.text('Avançar'));
-//    await tester.pumpAndSettle();
-
-    // Verifica se o terceiro título está visível
- //   expect(find.text('Como funciona?'), findsOneWidget);
-
-    // Verifica se o botão 'Avançar' desapareceu no último passo
-//    expect(find.text('Avançar'), findsNothing);
-
-    // Verifica se o botão 'Começar' aparece no último passo
-//    expect(find.text('Começar'), findsOneWidget);
-
-    // Clica no botão 'Começar' e aguarda
-//    await tester.tap(find.text('Começar'));
-//    await tester.pumpAndSettle();
-//  });
